refactor: introduce startup probe and decouple readiness from leadership
- Added /startup endpoint with startup atomic flag to gate initialization
- Orkestra now becomes Ready once servers are listening (not when leader elected)
- Followers are always Ready, ensuring stable webhook endpoints and multi‑Pod availability
- Leadership election moved to post-start hook after Orkestra is fully initialized
- Fixed circular dependency: readiness → endpoints → webhook → informers → workers → readiness
- Prevented premature finalizer removal during leadership transitions
- Clarified separation of concerns:
    * Orkestra = runtime + probes + servers
    * Konductor = elected leader
    * Kordinator = dependency orchestrator
- Improved reliability of conversion webhooks and informer startup
- Stabilized multi‑Pod orchestration and leader failover behavior
