+++ charts/orkestra/values.yaml
@@ -42,6 +42,14 @@ runtime:
     watchNamespace: ""
 
   # Probes
+  startupProbe:
+    httpGet:
+      path: /startup
+      port: health
+    initialDelaySeconds: 5
+    periodSeconds: 10
+    failureThreshold: 30
+
   livenessProbe:
     httpGet:
       path: /controlcenter/health
