+++ cmd/internal/banner.go
@@ -29,6 +29,7 @@ func printBanner(kfg *orkestraKfg, konductor string) {
 
 	// Endpoints
 	fmt.Println("Orkestra Endpoints:")
+	fmt.Printf("- Startup:   %s/startup%s\n", utils.ColorGreen, utils.ColorReset)
 	fmt.Printf("- Health:   %s/health%s\n", utils.ColorGreen, utils.ColorReset)
 	fmt.Printf("- Ready:    %s/ready%s\n", utils.ColorGreen, utils.ColorReset)
 	fmt.Printf("- Metrics:  %s/metrics%s\n", utils.ColorGreen, utils.ColorReset)
