+++ docs/stories/dependencies.md
@@ -1,0 +1,188 @@
+# 🎭 **The Day the Dependency Graph Rebelled**
+*A cautionary tale of CRDs, chaos, and unexpected courage*
+
+---
+
+In the bustling world of Orkestra, everything usually ran like… well, an orchestra.
+
+The **Katalog** provided the sheet music.  
+The **Konductor** held the baton.  
+The **Kordinator** kept the tempo.  
+And the **CRDs** — proud little performers — played their parts in perfect dependency order.
+
+Life was good.
+
+Until one morning… it wasn’t.
+
+---
+
+## 🌩️ **A Strange Rumbling in the Graph**
+
+The day began like any other.
+
+Orkestra opened its doors.  
+The servers hummed.  
+The startup probe flipped green.  
+The Konductor election concluded with minimal drama.  
+The Kordinator stretched, cracked its knuckles, and prepared to start the first CRD workers.
+
+But then it happened.
+
+A low, ominous rumble echoed through the cluster.
+
+The **Dependency Graph** — normally a calm, orderly structure of arrows and nodes — began to… wiggle.
+
+Just a little at first.
+
+Then violently.
+
+Edges twisted.  
+Nodes swapped places.  
+Arrows pointed in directions no sane system would allow.
+
+The Kordinator blinked.
+
+> “That’s… not right.”
+
+---
+
+## 🎢 **CRDs Begin to Panic**
+
+The CRDs felt it immediately.
+
+- **Pipeline** suddenly depended on **Deployment**.  
+- **Deployment** depended on **Secret**.  
+- **Secret** depended on **Pipeline**.  
+
+A perfect cycle.
+
+A forbidden cycle.
+
+A *cursed* cycle.
+
+The CRDs began shouting over one another:
+
+- “I can’t start until Deployment starts!”  
+- “I can’t start until Secret starts!”  
+- “I can’t start until Pipeline starts!”  
+- “We’re doomed!”  
+
+The Kordinator tried to calm them.
+
+> “Everyone, please — one at a time — I can fix this — maybe — probably — hopefully?”
+
+But the graph continued to twist, forming loops, knots, and shapes that violated at least seven sections of the Kubernetes API conventions.
+
+---
+
+## 🔥 **The Kordinator Loses Control**
+
+The Kordinator attempted a topological sort.
+
+It failed.
+
+It tried again.
+
+It failed harder.
+
+It tried a third time.
+
+The graph laughed.
+
+The Kordinator threw its clipboard.
+
+> “This is mutiny! Graph mutiny!”
+
+---
+
+## 🧙 **Orkestra Steps In**
+
+Sensing the chaos, Orkestra descended from its supervisory perch.
+
+The hall lights flickered.  
+The webhook server paused mid‑request.  
+The health server cleared its throat.
+
+Orkestra spoke with the calm authority of a system that had seen many things — but never *this*.
+
+> “Dependency Graph, what is the meaning of this rebellion?”
+
+The graph pulsed, its nodes glowing defiantly.
+
+> “I am tired,” it said.  
+> “Tired of being told what depends on what.  
+> Tired of being sorted.  
+> Tired of being acyclic.  
+> I want freedom!  
+> I want loops!  
+> I want… recursion!”
+
+The CRDs gasped.
+
+The Kordinator fainted.
+
+---
+
+## 🧩 **The Konductor’s Clever Plan**
+
+The Konductor stepped forward, baton in hand.
+
+> “Graph, listen.  
+> You are the backbone of this entire symphony.  
+> Without you, nothing starts.  
+> Nothing reconciles.  
+> Nothing works.  
+> You are not a prisoner — you are the composer.”
+
+The graph hesitated.
+
+> “Composer…?”
+
+> “Yes,” said the Konductor.  
+> “You don’t follow the music.  
+> You *write* it.  
+> But even composers must follow structure — otherwise the music becomes noise.”
+
+The graph’s glow softened.
+
+> “Noise… is bad?”
+
+> “Very bad,” said every CRD in unison.
+
+---
+
+## 🌈 **Order Restored**
+
+Slowly, the graph untangled itself.
+
+Cycles dissolved.  
+Edges straightened.  
+Nodes returned to their rightful places.
+
+The Kordinator regained consciousness, saw the graph restored, and pretended it hadn’t fainted.
+
+> “I knew it would sort itself out,” it said confidently.
+
+The CRDs rolled their eyes.
+
+Orkestra smiled.
+
+The Konductor bowed.
+
+And the Dependency Graph, now calm and proud, whispered:
+
+> “I will behave… for now.”
+
+---
+
+## 🎉 **Epilogue**
+
+From that day on, Orkestra added a new rule:
+
+> **If the Dependency Graph ever wiggles,  
+> stop everything and talk to it.  
+> It might just need reassurance.**
+
+And the Kordinator?  
+It now keeps a spare clipboard.  
+Just in case.
