+++ docs/stories/misising-finalizer.md
@@ -1,0 +1,204 @@
+# 🕵️‍♂️ **The Mystery of the Missing Finalizer**  
+*A suspenseful tale of CRDs, unfinished business, and a very confused Kordinator*
+
+---
+
+It was a peaceful morning in the Orkestra hall.
+
+The health server was humming.  
+The webhook server was polishing its certificates.  
+The Konductor was sipping a warm cup of logs.  
+The Kordinator was reviewing the day’s dependency graph with a sense of smug satisfaction.
+
+Everything was in order.
+
+Or so they thought.
+
+---
+
+## 🌫️ **A Strange Disappearance**
+
+The trouble began when a CRD named **Widget** approached the Kordinator with a trembling voice.
+
+> “Excuse me… I think something is wrong with my finalizer.”
+
+The Kordinator blinked.
+
+> “Your finalizer? What about it?”
+
+> “It’s… gone.”
+
+The Kordinator froze.
+
+Finalizers don’t just *go missing*.  
+They are sacred.  
+They are the guardians of cleanup.  
+They are the last line of defense against orphaned resources and cosmic disorder.
+
+The Kordinator leaned in.
+
+> “Show me.”
+
+Widget turned around, revealing its metadata.
+
+And indeed — the finalizer list was empty.
+
+Completely empty.
+
+The Kordinator gasped so loudly that three Deployments dropped their replicas.
+
+---
+
+## 🧩 **The Investigation Begins**
+
+The Kordinator summoned the CRDs.
+
+- Pipelines  
+- Secrets  
+- Deployments  
+- CronJobs  
+- ConfigMaps  
+- Even the shy little ServiceAccount  
+
+They gathered in a circle, whispering nervously.
+
+The Kordinator paced back and forth.
+
+> “Finalizers do not vanish.  
+> They are added deliberately.  
+> They are removed deliberately.  
+> Someone… or something… has tampered with the lifecycle.”
+
+The CRDs shuddered.
+
+Even the Dependency Graph wiggled anxiously.
+
+---
+
+## 🔍 **Clues in the Cache**
+
+The Kordinator inspected the informer cache.
+
+- No deletion timestamp  
+- No reconcile errors  
+- No PATCH events  
+- No suspicious logs  
+- No signs of a rogue controller  
+
+Everything looked normal.
+
+Too normal.
+
+The Kordinator narrowed its eyes.
+
+> “This smells like a race condition.”
+
+The CRDs gasped again.
+
+Race conditions were the stuff of legends — whispered about in dark YAML corners, feared by all, understood by none.
+
+---
+
+## 🧙 **Orkestra Arrives**
+
+Sensing the rising panic, Orkestra descended from its supervisory perch.
+
+The hall quieted.
+
+Orkestra spoke gently.
+
+> “Tell me what happened.”
+
+Widget stepped forward.
+
+> “I was being deleted.  
+> I felt the finalizer doing its job.  
+> The reconciler was preparing the cleanup.  
+> And then… everything went dark.  
+> When I woke up, the finalizer was gone — but the cleanup never happened.”
+
+Orkestra nodded slowly.
+
+> “A half‑completed deletion.  
+> A finalizer removed too early.  
+> A reconcile loop interrupted.”
+
+The Kordinator gulped.
+
+> “Interrupted… by what?”
+
+Orkestra turned toward the Dependency Graph.
+
+The graph wiggled nervously.
+
+---
+
+## ⚡ **The Truth Revealed**
+
+Orkestra spoke:
+
+> “The Konductor election.”
+
+The room fell silent.
+
+Orkestra continued:
+
+> “During Widget’s deletion, the leader lost its lease.  
+> A new Konductor was elected.  
+> The old leader stopped mid‑reconcile.  
+> The new leader picked up the object — but saw no finalizer.  
+> It assumed cleanup was done.”
+
+The CRDs gasped for the third time.
+
+The Kordinator slapped its forehead.
+
+> “Of course!  
+> The finalizer was removed *before* the cleanup finished.  
+> A classic premature‑finalizer‑removal scenario!”
+
+Widget sniffled.
+
+> “So… it wasn’t my fault?”
+
+> “No,” Orkestra said kindly.  
+> “You were a victim of leadership transition.”
+
+---
+
+## 🛠️ **The Fix**
+
+Orkestra raised its hand.
+
+A new rule appeared in glowing letters:
+
+> **“Finalizers must only be removed after cleanup is fully complete —  
+> and leadership transitions must never interrupt the process.”**
+
+The Kordinator nodded vigorously.
+
+> “I will enforce this.  
+> No worker will remove a finalizer until the cleanup is confirmed.  
+> And if leadership changes mid‑reconcile, the new leader will resume cleanup instead of skipping it.”
+
+Widget sighed with relief.
+
+The CRDs cheered.
+
+The Dependency Graph wiggled approvingly.
+
+---
+
+## 🎉 **Epilogue**
+
+From that day on:
+
+- No finalizer ever went missing again  
+- Cleanup was always completed  
+- Leadership transitions became graceful  
+- The Kordinator slept better at night  
+- And Widget lived happily ever after (until it was deleted properly)
+
+The mystery was solved.
+
+But the legend of the Missing Finalizer lived on — a reminder that even in a perfectly orchestrated system, timing is everything.
