+++ docs/stories/orkestra.md
@@ -1,0 +1,131 @@
+# 🎻 The Tale of Orkestra, the Konductor, and the Kordinator
+*A gentle saga of roles, responsibilities, and the dance of distributed systems*
+
+---
+
+In a quiet corner of the Kubernetes kingdom lived a grand ensemble known as **Orkestra**.  
+Not a single musician, but the entire symphony hall itself — the stage, the lights, the doors, the ushers, the ticket scanners, the whole infrastructure that makes music possible.
+
+Orkestra didn’t play instruments.  
+Orkestra didn’t conduct.  
+Orkestra didn’t decide who should play what.
+
+But Orkestra had one sacred duty:
+
+> **Make sure the show can go on.**
+
+So every morning, Orkestra would wake up early, stretch its HTTP servers, warm up its TLS certificates, polish its `/health`, `/ready`, and `/startup` signs, and open its doors to the Kubernetes kingdom.
+
+Only when everything was humming — the lights on, the doors open, the servers listening — would Orkestra flip its **startup flag** and announce:
+
+> “The hall is open. The show may begin.”
+
+---
+
+## 🎼 **Enter the Konductor**
+
+Once Orkestra was ready, it would ring a ceremonial bell.
+
+From across the cluster, Pods would perk up.  
+Each one dreamed of holding the baton.  
+Each one hoped to be chosen.
+
+This was the **Konductor Election** — a fair and ancient ritual.
+
+Pods would gather, whispering:
+
+- “Will it be me today?”
+- “I think I have a good chance.”
+- “I’ve been practicing my leadership loops.”
+
+But only one could win.
+
+When the election concluded, Orkestra would proudly announce:
+
+> “Behold! The Konductor has been chosen!”
+
+And the winner would step forward, bow gracefully, and take the baton.
+
+---
+
+## 🎺 **The Kordinator Awakens**
+
+Now, the Konductor didn’t make music either.  
+The Konductor’s job was to awaken the **Kordinator** — a brilliant, slightly overworked maestro responsible for the *actual orchestration*.
+
+The Kordinator knew every CRD by name.  
+It knew who depended on whom.  
+It knew which instruments (workers) needed to start first, and which ones must wait their turn.
+
+It would unfurl the **dependency graph**, tap the podium, and declare:
+
+> “Pipelines, you start first.  
+> Deployments, wait for your cue.  
+> CronJobs, hold your horses.  
+> Everyone will get their moment.”
+
+And so the great symphony of reconciliation began.
+
+---
+
+## 🎻 **Followers and Leaders**
+
+Meanwhile, the other Pods — the ones who didn’t win the election — didn’t sulk.
+
+They were **followers**, but proud ones.
+
+They kept the hall running:
+
+- serving webhooks  
+- answering health checks  
+- providing metrics  
+- staying ready in case the leader faltered  
+
+Because in Orkestra, leadership was a **role**, not an identity.
+
+If the Konductor ever dropped the baton — perhaps due to a node eviction or a surprise SIGTERM — Orkestra would simply ring the bell again.
+
+And another Pod would rise.
+
+---
+
+## 🎤 **The Grand Insight**
+
+One day, Orkestra realized something profound:
+
+> “Kubernetes doesn’t know about the Kordinator.  
+> Kubernetes doesn’t care about the Konductor.  
+> Kubernetes only sees *me*.”
+
+And with that, Orkestra understood:
+
+- **Readiness** is *Orkestra’s* responsibility  
+- **Leadership** is *Konductor’s* responsibility  
+- **Dependency orchestration** is *Kordinator’s* responsibility  
+
+Three layers.  
+Three roles.  
+One harmonious system.
+
+---
+
+## 🎇 **And so the show goes on**
+
+Every day:
+
+- Orkestra opens the hall  
+- The Konductor is chosen  
+- The Kordinator orchestrates  
+- The CRDs perform  
+- The cluster applauds  
+
+And Kubernetes, blissfully unaware of the intricate dance inside, simply sees:
+
+- a healthy Pod  
+- a ready Pod  
+- a stable webhook  
+- a reliable controller  
+
+But you — the architect — know the truth.
+
+You built a symphony.
