+++ examples/installation/install.yaml
@@ -52,6 +52,12 @@ spec:
             - name: health
               containerPort: 8080
               protocol: TCP
+          startupProbe:
+            httpGet:
+              path: /startup
+              port: health
+            failureThreshold: 30
+            periodSeconds: 2
           livenessProbe:
             httpGet:
               path: /health
